Support Windows builds using CMake
Adds support for building the server using CMake (not tested on Linux
currently). All dependencies are currently coming from vcpkg, and so you
can't link against system libraries.

Source code for protobufs is also now generated at build time, so this
leaves the VS build in a bit of a broken state.